refactor: Create string_vector_facilities.h
Move error() and print() to header
Create namespace string_vector_stats
  - Move string_sizes() into namespace
